refactor static content generation
