+++ apps/www/components/Nav/DevelopersDropdown.tsx
@@ -2,10 +2,11 @@ import { ChevronRight } from 'lucide-react'
 import Link from 'next/link'
 
 import { data as DevelopersData } from 'data/Developers'
-import blogPosts from '~/.contentlayer/generated/LatestBlogPost/_index.json' with { type: 'json' }
-import openJobPositions from '~/.contentlayer/generated/CareersCount/_index.json' with { type: 'json' }
+import {
+  jobsCount,
+  latestBlogPosts,
+} from '~/.contentlayer/generated/staticContent/_index.json' with { type: 'json' }
 import { Badge } from 'ui'
-const jobsCount = openJobPositions.jobsCount
 
 type LinkProps = {
   text: string
@@ -33,9 +34,9 @@ const DevelopersDropdown = () => (
                   {Icon && <Icon size={16} strokeWidth={1.3} />}
                   <span>{link.text}</span>
                   {link.text === 'Careers' && jobsCount > 0 && (
-                    <Badge className="p-0 !bg-surface-200 w-[22px] h-[22px] border-none text-xs flex items-center justify-center text-foreground-lighter">
-                      {openJobPositions.jobsCount}
-                    </Badge>
+                    <span className="text-xs flex items-center justify-center text-foreground-lighter">
+                      ({jobsCount})
+                    </span>
                   )}
                   <ChevronRight
                     strokeWidth={2}
@@ -58,7 +59,7 @@ const DevelopersDropdown = () => (
           <ChevronRight className="h-3 w-3 transition-transform will-change-transform -translate-x-1 group-hover:translate-x-0" />
         </Link>
         <ul className="flex flex-col gap-5">
-          {blogPosts?.map((post) => (
+          {latestBlogPosts?.map((post) => (
             <li key={post.title}>
               <Link
                 href={post.url}
