+++ apps/www/package.json
@@ -12,7 +12,7 @@
     "lint": "next lint",
     "clean": "rimraf node_modules",
     "typecheck": "pnpm run content:build && tsc --noEmit",
-    "content:build": "contentlayer2 build && node scripts/generateLatestBlogPosts.mjs && node scripts/generateCareersLength.mjs",
+    "content:build": "contentlayer2 build && node scripts/generateStaticContent.mjs",
     "prettier": "prettier --write \"./{pages,components,lib,stores,styles,tests}/**/*.{ts,tsx,md,js,jsx,json}\"",
     "postbuild": "node ./internals/generate-sitemap.mjs"
   },
