+++ apps/www/scripts/generateStaticContent.mjs
@@ -1,6 +1,8 @@
 import { promises as fs } from 'fs'
 import path from 'path'
 import { fileURLToPath } from 'url'
+
+// Latest Blog Posts
 import { allBlogPosts } from '../.contentlayer/generated/index.mjs'
 
 /**
@@ -27,8 +29,25 @@ const latestBlogPosts = allBlogPosts
   .slice(0, 2)
   .map(({ title, url, description }) => ({ title, url, description }))
 
+// Careers Jobs count
+const getCareerCount = async () => {
+  const job_res = await fetch('https://api.ashbyhq.com/posting-api/job-board/supabase')
+  const job_data = await job_res.json()
+
+  return job_data.jobs.length
+}
+
+let careersCount = 0
+
+try {
+  careersCount = await getCareerCount()
+} catch (error) {
+  throw error
+}
+
+// Create folder for static content
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
-const folderPath = path.join(__dirname, '../.contentlayer/generated/LatestBlogPost')
+const folderPath = path.join(__dirname, '../.contentlayer/generated/staticContent')
 try {
   await fs.mkdir(folderPath, { recursive: true })
 } catch (error) {
@@ -38,5 +57,10 @@ try {
   // Folder already exists, continue silently
 }
 
-const filePath = path.join(__dirname, '../.contentlayer/generated/LatestBlogPost/_index.json')
-await fs.writeFile(filePath, JSON.stringify(latestBlogPosts), 'utf8')
+// Write static content to file
+const filePath = path.join(__dirname, '../.contentlayer/generated/staticContent/_index.json')
+await fs.writeFile(
+  filePath,
+  JSON.stringify({ latestBlogPosts: latestBlogPosts, jobsCount: careersCount }),
+  'utf8'
+)
